demo/otel-demo-app/requirements.txt
flask==3.0.3
gunicorn==22.0.0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-flask==0.47b0
opentelemetry-instrumentation-requests==0.47b0
